Retry gas estimation
If a transaction to be sent fails, the error we'll actually get is
an 'always failing transaction' on the estimateGas we sent first.
However, we want to retry some transactions, as they may depend on
a previous tx being mined, which may have not propagated yet.
To handle these scenarios, we add 3 retries to gas estimation,
separated by a small sleep interval.

Fixes #192