Handle initialize function lookup on dependencies contracts
Parsing the linearized base contracts on contracts in dependencies
failed, since their absolute paths pointed to inexistent directories in
the local machine, as they were compiled somewhere else. Also, truffle
uses prefixes when compiling contracts from dependencies, but not when
compiling local contracts: this causes the absolute path not to match
between the ImportDirective and the compiled artifact.

To workaround this, we catch any errors on that lookup, and fall back to
matching on the ABI directly by number of args only. If we detect more
than one method, we fail.

To allow the user to pinpoint a method, now we accept a fully qualified
function name, like `myFunction(uint256,string)`. This shortcircuits the
initializer function lookup, and goes straight for the ABI, matching by
the specified argument types.