Fix storage layout parsing when there are complex types (#286)
Given a contract with a mapping or array (not struct), the check for
whether it contained structs or enums failed if there was no variable
with the same type as the value type of the mapping or array. For
instance, a contract with a variable of type `array(string)` and another
`string` would work, but a contract with just `array(string)` would
fail.