Check duplicated artifact names (#880)
If there are two contracts with the same name on the same project, truffle will compile them to the same artifact (ContractName.json), overwriting one of them. This causes some validations to fail, such as storage validations, since they expect a set of AST nodes that is not found due to the overwritten artifact. It also yields false negatives in other validations, such as no-delegatecalls, since it looks for all parent contracts, but it could end up loading an incorrect one (since the right one was overwritten).

This change walks through the set of ancestors of a given contract, and checks that the referenced AST ID for each of them is present among the set of imported artifacts. While this was doable by just trying to load the nodes that corresponded to the linearized base contracts, they only contain the IDs of the nodes, and not their friendly names. So we added a method to recursively walk the set of ancestors, keeping track of both id and name, so we can show a more descriptive message to the user:

```
Error while validating contract MyChildContract: Cannot find source data for contract MyContract (base contract of MyChildContract). This often happens because either:
- An incremental compilation step went wrong. Clear your build folder and recompile.
- There is more than one contract named MyContract in your project (including dependencies). Make sure all contracts have a unique name, and that you are not importing dependencies with duplicated contract names (for example, openzeppelin-eth and openzeppelin-solidity).
```

Fixes #389