Do not write invalid zos.json if compiling before init
Running zos compile in a project without zos.json would create a zos.json file without name or version, which caused the project to crash later.

There were two options to fix this:
1- Refuse to compile before init
2- Compile anyway without persisting compiler flags

I went with (2), to make it easier to use zos as a standalone compiler, without needing to init a project wherever the user wants to run it.