Rewrite resolver engine gatherer
All cases are now manually handled without a dependency to resolver-engine. Support for relative imports not starting with '.' is removed, unless they are relative to the project root, like 'contracts/foo/bar.sol'. Note that these are not supported for dependencies, only for project contracts.

A new compiler setting, workingDir, is added. All relative imports not starting with '.' are resolved on this directory.