Rework compiler imports (#1411)
* Rearrange mock projects to compile

- Move mock project to compile to test/mocks
- Remove stdlib-with-deps as it is already covered in mock-project-to-compile
- Add package.json where missing
- Remove invalid import statements

* Rewrite resolver engine gatherer

All cases are now manually handled without a dependency to resolver-engine. Support for relative imports not starting with '.' is removed, unless they are relative to the project root, like 'contracts/foo/bar.sol'. Note that these are not supported for dependencies, only for project contracts.

A new compiler setting, workingDir, is added. All relative imports not starting with '.' are resolved on this directory.

* Remove resolver-engine

* Rename ResolverEngineGatherer

* Use glob for searching for project solidity files

* Lint

* Handle projects with repeated contract names

* Update packages/cli/src/models/compiler/solidity/SourcesGatherer.ts

Co-Authored-By: Igor Yalovoy <to@ylv.io>

Co-authored-by: Igor Yalovoy <to@ylv.io>