Support solidity 0.6 in Initializable contract (#1413)
* Support solidity 0.6 in Initializable contract

Modifies Initializable so it is compatible with Solidity 0.5 and 0.6. Adds a set of tests that run on the 0.6 version to verify its behaviour (note that the 0.6 compiled contracts are committed into source control to simplify the setup).

* Move mock packages into their own workspaces

Both mock-solc and mock-dependency where being copied instead of symlinked into packages/lib by yarn. Even worse, they were not re-copied if they were modified, requiring to rm-rf them, or update the cache in the CI.

By defining mock-solc and mock-dependency as workspaces, yarn will symlink them from where they are requested, instead of copying them. This required moving them outside the packages/lib workspace, as yarn does not support nested workspaces.

* Add missing compiled artifacts